+++ js/data.js
@@ -1,0 +1,48 @@
+const TITLES = [
+  'Дворец',
+  'Квартира',
+  'Дом',
+  'Бунгало',
+  'Отель',
+];
+const TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+const CHECKIN_OUT = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+const DESCRIPTIONS = [
+  'Уютная квартира',
+  'С видом на улицу',
+  'Солнечная сторона',
+  'Удобная парковка',
+  'Круглосуточное заселение',
+  'Можно с животными',
+];
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+const MIN_LAT = 35.65000;
+const MAX_LAT = 35.70000;
+const MIN_LNG = 139.70000;
+const MAX_LNG = 139.80000;
+const OBJECT_COUNT = 10;
+
+
+export {MIN_LAT, MAX_LAT, MIN_LNG, MAX_LNG, OBJECT_COUNT, PHOTOS, DESCRIPTIONS, FEATURES, CHECKIN_OUT, TYPES, TITLES};
