+++ js/main.js
@@ -1,68 +1,5 @@
-function getRandomPositiveFloat (a, b, digits = 5) {
-
-  const lower = Math.min(Math.abs(a), Math.abs(b));
-  const upper = Math.max(Math.abs(a), Math.abs(b));
-  const result = Math.random() * (upper - lower) + lower;
-
-  return result.toFixed(digits);
-}
-
-function getRandomPositiveInteger (a, b) {
-  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
-  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-}
-
-const TITLES = [
-  'Дворец',
-  'Квартира',
-  'Дом',
-  'Бунгало',
-  'Отель',
-];
-const TYPES = [
-  'palace',
-  'flat',
-  'house',
-  'bungalow',
-  'hotel',
-];
-const CHECKIN_OUT = [
-  '12:00',
-  '13:00',
-  '14:00',
-];
-const FEATURES = [
-  'wifi',
-  'dishwasher',
-  'parking',
-  'washer',
-  'elevator',
-  'conditioner',
-];
-const DESCRIPTIONS = [
-  'Уютная квартира',
-  'С видом на улицу',
-  'Солнечная сторона',
-  'Удобная парковка',
-  'Круглосуточное заселение',
-  'Можно с животными',
-];
-const PHOTOS = [
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
-];
-const MIN_LAT = 35.65000;
-const MAX_LAT = 35.70000;
-const MIN_LNG = 139.70000;
-const MAX_LNG = 139.80000;
-// eslint-disable-next-line no-unused-vars
-const NUMBER_OBJECTS = 10;
-
-const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
-const getArrayRandomLength = (array) => array.slice(getRandomPositiveInteger(1, array.length - 1));
+import {MIN_LAT, MAX_LAT, MIN_LNG, MAX_LNG, OBJECT_COUNT, PHOTOS, DESCRIPTIONS, FEATURES, CHECKIN_OUT, TYPES, TITLES} from './data.js';
+import {getRandomArrayElement, getArrayRandomLength, getRandomPositiveInteger, getRandomPositiveFloat} from './util.js';
 
 const getLocation = () => {
   getRandomPositiveFloat(MIN_LAT, MAX_LAT);
@@ -87,7 +24,6 @@ const createOffer = () => {
     location: location,
   };
 };
-// eslint-disable-next-line no-undef
+
 const createOffers = Array.from({length: OBJECT_COUNT}, createOffer);
 createOffers();
-
