+++ js/util.js
@@ -1,0 +1,19 @@
+function getRandomPositiveFloat (a, b, digits = 5) {
+  const lower = Math.min(Math.abs(a), Math.abs(b));
+  const upper = Math.max(Math.abs(a), Math.abs(b));
+  const result = Math.random() * (upper - lower) + lower;
+  return result.toFixed(digits);
+}
+
+function getRandomPositiveInteger (a, b) {
+  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+}
+
+const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+
+const getArrayRandomLength = (array) => array.slice(getRandomPositiveInteger(1, array.length - 1));
+
+export {getRandomArrayElement, getArrayRandomLength, getRandomPositiveInteger, getRandomPositiveFloat};
